Split E2E testing (#19)
## Why is this change necessary?
Lots of flakiness when trying to run dockerized and regular server E2E
tests as part of the same suite (conflicts with ports, failed builds...)


 ## How does this change address the issue?
Splits it to run as two separate suites. Same tests, just docker-compose
vs not. Removes the `$fetch` test of static HTML, which seems to provide
minimal value.


 ## What side effects does this change have?
Extra command to run tests


 ## How is this change tested?
In CI and locally in downstream repos

+++ template/frontend/tests/e2e/index.spec.ts
@@ -1,45 +1,32 @@
 import { test, expect, describe, beforeAll, afterAll } from "vitest";
-import { setup, $fetch, createPage, url } from "@nuxt/test-utils/e2e";
+import { setup, createPage, url } from "@nuxt/test-utils/e2e";
 import { execSync } from "child_process";
-describe("Index page", async () => {
-  describe("regular nuxt build", async () => {
-    await setup();
-    describe("Static tests of rendered page", async () => {
-      test("Page displays Hello World", async () => {
-        const html = await $fetch("/");
-        expect(html).toContain("Hello World");
-      });
-    });
-    describe("Full Browser tests", async () => {
-      // TODO: figure out away to display the chromium version so it will show up in logs
-      test("Page displays Hello World", async () => {
-        const page = await createPage();
-        await page.goto(url("/"), { waitUntil: "hydration" });
-        const text = await page.textContent("div");
 
-        expect(text).toContain("Hello World");
-      });
-    });
-  });
-  describe("docker build", async () => {
-    beforeAll(async () => {
+describe("Index page", async () => {
+  if (!process.env.USE_DOCKER_COMPOSE_FOR_VITEST_E2E) {
+    // the `setup()` function doesn't seem to work well when trying to run tests in the same suite with different hosts, so using an envvar to run the tests either with docker-compose or without
+    await setup();
+  }
+  beforeAll(async () => {
+    if (process.env.USE_DOCKER_COMPOSE_FOR_VITEST_E2E) {
       console.log("Starting docker-compose...");
       execSync("docker compose --file=../docker-compose.yaml up --detach --force-recreate --renew-anon-volumes", {
         stdio: "inherit",
       });
       await setup({ host: "http://localhost:3000" });
-    }, 240 * 1000); // increase timeout in case image needs to be built
-    afterAll(() => {
+    }
+  }, 240 * 1000); // increase timeout in case image needs to be built
+  afterAll(() => {
+    if (process.env.USE_DOCKER_COMPOSE_FOR_VITEST_E2E) {
       console.log("Stopping docker-compose...");
       execSync("docker compose --file=../docker-compose.yaml down", { stdio: "inherit" });
-    });
+    }
+  });
+  test("Page displays Hello World", async () => {
+    const page = await createPage();
+    await page.goto(url("/"), { waitUntil: "hydration" });
+    const text = await page.textContent("div");
 
-    test("Page displays Hello World", async () => {
-      const page = await createPage();
-      await page.goto(url("/"), { waitUntil: "hydration" });
-      const text = await page.textContent("div");
-
-      expect(text).toContain("Hello World");
-    });
+    expect(text).toContain("Hello World");
   });
 });
